% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_mas5}
\alias{filter_mas5}
\title{Filter ExpressionSet by MAS5 presence}
\usage{
filter_mas5(eset, min_present = 10)
}
\arguments{
\item{eset}{An ExpressionSet object.}

\item{min_present}{Minimum number of samples (out of 48) where a probe
must be 'P' to be kept. Default is 10.}
}
\value{
A filtered ExpressionSet object.
}
\description{
Filters probes based on the number of samples where they were called 'Present'
(MAS5 algorithm).
}
\examples{
data(eset)
# Filter to keep probes present in at least 12 samples
filtered_eset <- filter_mas5(eset, min_present = 12)
}
