% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_boxplots}
\alias{plot_boxplots}
\title{Plot Overlaid Boxplots for all samples}
\usage{
plot_boxplots(eset, title = "Sample Expression Distribution")
}
\arguments{
\item{eset}{An ExpressionSet object.}

\item{title}{Plot title.}
}
\description{
Plot Overlaid Boxplots for all samples
}
