% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_densities}
\alias{plot_densities}
\title{Plot Overlaid Density for all samples}
\usage{
plot_densities(eset, title = "Sample Density Distributions")
}
\arguments{
\item{eset}{An ExpressionSet object.}

\item{title}{Plot title.}
}
\description{
Plot Overlaid Density for all samples
}
